add uint32 of calcul of the pulse

+++ TIMER_OC_PWM_1/Src/main.c
@@ -203,26 +203,26 @@ void UART2_Init(void)
 	 tim2PWM_Config.OCPolarity = TIM_OCPOLARITY_HIGH;
 
 
-	tim2PWM_Config.Pulse =  (htimer2.Init.Period * 25 ) /100;
+	tim2PWM_Config.Pulse =  (uint32_t)(htimer2.Init.Period * 25 ) /100;
 
 	 if( HAL_TIM_PWM_ConfigChannel(&htimer2,&tim2PWM_Config,TIM_CHANNEL_1) != HAL_OK)
 	 {
 		 Error_handler();
 	 }
 
-	 tim2PWM_Config.Pulse =  (htimer2.Init.Period * 45 ) /100;
+	 tim2PWM_Config.Pulse =  (uint32_t)(htimer2.Init.Period * 45 ) /100;
 	 if( HAL_TIM_PWM_ConfigChannel(&htimer2,&tim2PWM_Config,TIM_CHANNEL_2) != HAL_OK)
 	 {
 		 Error_handler();
 	 }
 
-	 tim2PWM_Config.Pulse =  (htimer2.Init.Period * 75 ) /100;
+	 tim2PWM_Config.Pulse =  (uint32_t)(htimer2.Init.Period * 75 ) /100;
 	 if( HAL_TIM_PWM_ConfigChannel(&htimer2,&tim2PWM_Config,TIM_CHANNEL_3) != HAL_OK)
 	 {
 		 Error_handler();
 	 }
 
-	 tim2PWM_Config.Pulse =  (htimer2.Init.Period * 95 ) /100;
+	 tim2PWM_Config.Pulse =  (uint32_t)(htimer2.Init.Period * 95 ) /100;
 	 if( HAL_TIM_PWM_ConfigChannel(&htimer2,&tim2PWM_Config,TIM_CHANNEL_4) != HAL_OK)
 	 {
 		 Error_handler();
